Move WindhshaftCLient to its own folder
